Split walking and segmenting a path into its own class so it can be used by other algorithms, not just visualisation
