+++ src/Mod/Path/Gui/ViewProviderPath.cpp
@@ -49,6 +49,7 @@
 
 #include <Mod/Path/App/FeaturePath.h>
 #include <Mod/Path/App/Path.h>
+#include <Mod/Path/App/PathSegmentWalker.h>
 #include <App/Application.h>
 #include <App/Document.h>
 #include <Base/FileInfo.h>
@@ -60,18 +61,6 @@
 #include <Gui/SoFCBoundingBox.h>
 #include <Gui/SoAxisCrossKit.h>
 #include <Gui/SoFCUnifiedSelection.h>
-
-
-#define ARC_MIN_SEGMENTS   20.0  // minimum # segments to interpolate an arc
-
-#ifndef M_PI
-    #define M_PI 3.14159265358979323846
-    #define M_PI    3.14159265358979323846 /* pi */
-#endif
-
-#ifndef M_PI_2
-    #define M_PI_2  1.57079632679489661923 /* pi/2 */
-#endif
 
 using namespace Gui;
 using namespace PathGui;
@@ -463,19 +452,146 @@ void ViewProviderPath::hideSelection() {
     pcArrowSwitch->whichChild = -1;
 }
 
-Base::Vector3d compensateRotation(const Base::Vector3d &pt, const Base::Rotation &rot, const Base::Vector3d &center)
+class VisualPathSegmentVisitor
+: public PathSegmentVisitor
 {
-    Base::Vector3d ptRotated;
-    rot.multVec(pt - center, ptRotated);
-    return ptRotated + center;
-}
+public:
+    VisualPathSegmentVisitor(
+        const Toolpath &tp,
+        SoCoordinate3 *pcLineCoords_,
+        SoCoordinate3 *pcMarkerCoords_,
+        std::vector<int> &command2Edge_,
+        std::deque<int> &edge2Command_,
+        std::deque<int> &edgeIndices_,
+        std::vector<int> &colorindex_,
+        std::deque<Base::Vector3d> &points_,
+        std::deque<Base::Vector3d> &markers_)
+    : pcLineCoords(pcLineCoords_)
+    , pcMarkerCoords(pcMarkerCoords_)
+    , command2Edge(command2Edge_)
+    , edge2Command(edge2Command_)
+    , edgeIndices(edgeIndices_)
+    , colorindex(colorindex_)
+    , points(points_)
+    , markers(markers_)
+    {
+        pcLineCoords->point.deleteValues(0);
+        pcMarkerCoords->point.deleteValues(0);
 
-Base::Rotation yawPitchRoll(double a, double b, double c)
-{
-    Base::Rotation rot; 
-    rot.setYawPitchRoll(-c, -b, -a);
-    return rot;
-}
+        command2Edge.clear();
+        edge2Command.clear();
+        edgeIndices.clear();
+
+        colorindex.clear();
+
+        command2Edge.resize(tp.getSize(),-1);
+    }
+
+    virtual void setup(const Base::Vector3d &last)
+    {
+        points.push_back(last);
+        markers.push_back(last);
+    }
+
+    virtual void g0(int id, const Base::Vector3d &last, const Base::Vector3d &next, const std::deque<Base::Vector3d> &pts)
+    {
+        (void)last;
+        gx(id, &next, pts, 0);
+    }
+
+    virtual void g1(int id, const Base::Vector3d &last, const Base::Vector3d &next, const std::deque<Base::Vector3d> &pts)
+    {
+        (void)last;
+        gx(id, &next, pts, 1);
+    }
+
+    virtual void g23(int id, const Base::Vector3d &last, const Base::Vector3d &next, const std::deque<Base::Vector3d> &pts, const Base::Vector3d &center)
+    {
+        (void)last;
+        gx(id, &next, pts, 1);
+        markers.push_back(center);
+    }
+
+    virtual void g8x(int id, const Base::Vector3d &last, const Base::Vector3d &next, const std::deque<Base::Vector3d> &pts,
+                     const std::deque<Base::Vector3d> &p, const std::deque<Base::Vector3d> &q)
+    {
+        (void)last;
+
+        gx(id, NULL, pts, 0);
+
+        points.push_back(p[0]);
+        markers.push_back(p[0]);
+        colorindex.push_back(0);
+
+        points.push_back(p[1]);
+        markers.push_back(p[1]);
+        colorindex.push_back(0);
+
+        points.push_back(next);
+        markers.push_back(next);
+        colorindex.push_back(1);
+
+        for (std::deque<Base::Vector3d>::const_iterator it=q.begin(); q.end() != it; ++it) {
+            markers.push_back(*it);
+        }
+
+        points.push_back(p[2]);
+        markers.push_back(p[2]);
+        colorindex.push_back(0);
+
+        pushCommand(id);
+    }
+
+    virtual void g38(int id, const Base::Vector3d &last, const Base::Vector3d &next)
+    {
+      Base::Vector3d p1(next.x,next.y,last.z);
+      points.push_back(p1);
+      colorindex.push_back(0);
+
+      points.push_back(next);
+      colorindex.push_back(2);
+
+      Base::Vector3d p3(next.x,next.y,last.z);
+      points.push_back(p3);
+      colorindex.push_back(0);
+
+      pushCommand(id);
+    }
+
+private:
+    SoCoordinate3 *pcLineCoords;
+    SoCoordinate3 *pcMarkerCoords;
+
+    std::vector<int> &command2Edge;
+    std::deque<int> &edge2Command;
+    std::deque<int> &edgeIndices;
+
+    std::vector<int> &colorindex;
+    std::deque<Base::Vector3d> &points;
+    std::deque<Base::Vector3d> &markers;
+
+    virtual void gx(int id, const Base::Vector3d *next, const std::deque<Base::Vector3d> &pts, int color)
+    {
+        for (std::deque<Base::Vector3d>::const_iterator it=pts.begin(); pts.end() != it; ++it) {
+          points.push_back(*it);
+          colorindex.push_back(color);
+        }
+
+        if (next != NULL) {
+            points.push_back(*next);
+            markers.push_back(*next);
+            colorindex.push_back(color);
+
+            pushCommand(id);
+        }
+    }
+
+    void pushCommand(int id) {
+      command2Edge[id] = edgeIndices.size();
+      edgeIndices.push_back(points.size());
+      edge2Command.push_back(id);
+    }
+};
 
 void ViewProviderPath::updateVisual(bool rebuild) {
 
@@ -486,282 +602,25 @@ void ViewProviderPath::updateVisual(bool rebuild) {
     pcLines->coordIndex.deleteValues(0);
 
     if(rebuild) {
-        pcLineCoords->point.deleteValues(0);
-        pcMarkerCoords->point.deleteValues(0);
-
-        command2Edge.clear();
-        edge2Command.clear();
-        edgeIndices.clear();
-
         Path::Feature* pcPathObj = static_cast<Path::Feature*>(pcObject);
         const Toolpath &tp = pcPathObj->Path.getValue();
-        if(tp.getSize()==0) {
-            return;
-        }
 
-        ParameterGrp::handle hGrp = App::GetApplication().GetParameterGroupByPath("User parameter:BaseApp/Preferences/Mod/Part");
-        float deviation = hGrp->GetFloat("MeshDeviation",0.2);
         std::deque<Base::Vector3d> points;
         std::deque<Base::Vector3d> markers;
 
-        Base::Vector3d rotCenter = tp.getCenter();
-        Base::Vector3d last(StartPosition.getValue());
-        Base::Rotation lrot;
-        double A = 0.0;
-        double B = 0.0;
-        double C = 0.0;
+        VisualPathSegmentVisitor collect(tp,
+            pcLineCoords,
+            pcMarkerCoords,
+            command2Edge,
+            edge2Command,
+            edgeIndices,
+            colorindex,
+            points,
+            markers);
 
-        colorindex.clear();
-        bool absolute = true;
-        bool absolutecenter = false;
+        PathSegmentWalker segments(tp);
+        segments.walk(collect, StartPosition.getValue());
 
-        // for mapping the coordinates to XY plane
-        double Base::Vector3d::*pz = &Base::Vector3d::z;
-
-        command2Edge.resize(tp.getSize(),-1);
-
-        points.push_back(last);
-        markers.push_back(last); // startpoint of path
-
-        for (unsigned int  i = 0; i < tp.getSize(); i++) {
-            const Path::Command &cmd = tp.getCommand(i);
-            const std::string &name = cmd.Name;
-            Base::Vector3d next = cmd.getPlacement().getPosition();
-            double a = A;
-            double b = B;
-            double c = C;
-
-            if (!absolute)
-                next = last + next;
-            if (!cmd.has("X")) next.x = last.x;
-            if (!cmd.has("Y")) next.y = last.y;
-            if (!cmd.has("Z")) next.z = last.z;
-            if ( cmd.has("A")) a = cmd.getValue("A");
-            if ( cmd.has("B")) b = cmd.getValue("B");
-            if ( cmd.has("C")) c = cmd.getValue("C");
-
-            Base::Rotation nrot = yawPitchRoll(a, b, c);
-
-            Base::Vector3d rnext = compensateRotation(next, nrot, rotCenter);
-
-            if ( (name == "G0") || (name == "G00") || (name == "G1") || (name == "G01") ) {
-                // straight line
-                int color = ((name == "G0") || (name == "G00")) ? 0 : 1;
-                if (nrot != lrot) {
-                    double amax = std::max(fmod(fabs(a - A), 360), std::max(fmod(fabs(b - B), 360), fmod(fabs(c - C), 360)));
-                    double angle = amax / 180 * M_PI;
-                    int segments = std::max(ARC_MIN_SEGMENTS, 3.0/(deviation/angle));
-
-                    double da = (a - A) / segments;
-                    double db = (b - B) / segments;
-                    double dc = (c - C) / segments;
-
-                    Base::Vector3d dnext = (next - last) / segments;
-
-                    for (int j = 1; j < segments; j++) {
-                        Base::Vector3d inter = last + dnext * j;
-
-                        Base::Rotation rot = yawPitchRoll(A + da*j, B + db*j, C + dc*j);
-                        Base::Vector3d rinter = compensateRotation(inter, rot, rotCenter);
-
-                        points.push_back(rinter);
-                        colorindex.push_back(color);
-                    }
-                }
-                points.push_back(rnext);
-                markers.push_back(rnext); // endpoint
-                colorindex.push_back(color); // std color
-
-                command2Edge[i] = edgeIndices.size();
-                edgeIndices.push_back(points.size());
-                edge2Command.push_back(i);
-
-                last = next;
-                A = a;
-                B = b;
-                C = c;
-                lrot = nrot;
-
-            } else if ( (name == "G2") || (name == "G02") || (name == "G3") || (name == "G03") ) {
-                // arc
-                Base::Vector3d norm;
-                Base::Vector3d center;
-
-                if ( (name == "G2") || (name == "G02") )
-                    norm.*pz = -1.0;
-                else
-                    norm.*pz = 1.0;
-
-                if (absolutecenter)
-                    center = cmd.getCenter();
-                else
-                    center = (last + cmd.getCenter());
-                Base::Vector3d next0(next);
-                next0.*pz = 0.0;
-                Base::Vector3d last0(last);
-                last0.*pz = 0.0;
-                Base::Vector3d center0(center);
-                center0.*pz = 0.0;
-                //double radius = (last - center).Length();
-                double angle = (next0 - center0).GetAngle(last0 - center0);
-                // GetAngle will always return the minor angle. Switch if needed
-                Base::Vector3d anorm = (last0 - center0) % (next0 - center0);
-                if (anorm.*pz < 0) {
-                    if(name == "G3" || name == "G03")
-                        angle = M_PI * 2 - angle;
-                } else if(anorm.*pz > 0) {
-                    if(name == "G2" || name == "G02")
-                        angle = M_PI * 2 - angle;
-                } else if (angle == 0)
-                    angle = M_PI * 2;
-
-                double amax = std::max(fmod(fabs(a - A), 360), std::max(fmod(fabs(b - B), 360), fmod(fabs(c - C), 360)));
-
-                int segments = std::max(ARC_MIN_SEGMENTS, 3.0/(deviation/std::max(angle, amax))); //we use a rather simple rule here, provisorily
-                double dZ = (next.*pz - last.*pz)/segments; //How far each segment will helix in Z
-
-                double dangle = angle/segments;
-                double da = (a - A) / segments;
-                double db = (b - B) / segments;
-                double dc = (c - C) / segments;
-
-                for (int j = 1; j < segments; j++) {
-                    Base::Vector3d inter;
-                    Base::Rotation rot(norm, dangle*j);
-                    rot.multVec((last0 - center0), inter);
-                    inter.*pz = last.*pz + dZ * j; //Enable displaying helices
-
-                    Base::Rotation arot = yawPitchRoll(A + da*j, B + db*j, C + dc*j);
-                    Base::Vector3d rinter = compensateRotation(center0 + inter, arot, rotCenter);
-
-                    points.push_back(rinter);
-                    colorindex.push_back(1);
-                }
-
-                points.push_back(rnext);
-                markers.push_back(rnext); // endpoint
-                markers.push_back(center); // add a marker at center too
-
-                colorindex.push_back(1);
-                command2Edge[i] = edgeIndices.size();
-                edgeIndices.push_back(points.size());
-                edge2Command.push_back(i);
-
-                last = next;
-                A = a;
-                B = b;
-                C = c;
-                lrot = nrot;
-
-            } else if (name == "G90") {
-                // absolute mode
-                absolute = true;
-
-            } else if (name == "G91") {
-                // relative mode
-                absolute = false;
-
-            } else if (name == "G90.1") {
-                // absolute mode
-                absolutecenter = true;
-
-            } else if (name == "G91.1") {
-                // relative mode
-                absolutecenter = false;
-
-            } else if ((name=="G81")||(name=="G82")||(name=="G83")||(name=="G84")||(name=="G85")||(name=="G86")||(name=="G89")){
-                // drill,tap,bore
-                double r = 0;
-                if (cmd.has("R"))
-                    r = cmd.getValue("R");
-
-                Base::Vector3d p1(next);
-                p1.*pz = last.*pz;
-
-                if (nrot != lrot) {
-                    double amax = std::max(fmod(fabs(a - A), 360), std::max(fmod(fabs(b - B), 360), fmod(fabs(c - C), 360)));
-                    double angle = amax / 180 * M_PI;
-                    int segments = std::max(ARC_MIN_SEGMENTS, 3.0/(deviation/angle));
-
-                    double da = (a - A) / segments;
-                    double db = (b - B) / segments;
-                    double dc = (c - C) / segments;
-
-                    Base::Vector3d dnext = (p1 - last) / segments;
-
-                    for (int j = 1; j < segments; j++) {
-                        Base::Vector3d inter = last + dnext * j;
-
-                        Base::Rotation rot = yawPitchRoll(A + da*j, B + db*j, C + dc*j);
-                        Base::Vector3d rinter = compensateRotation(inter, rot, rotCenter);
-
-                        points.push_back(rinter);
-                        colorindex.push_back(0);
-                    }
-                }
-
-                Base::Vector3d p1r = compensateRotation(p1, nrot, rotCenter);
-                points.push_back(p1r);
-                markers.push_back(p1r);
-                colorindex.push_back(0);
-                Base::Vector3d p2(next);
-                p2.*pz = r;
-                Base::Vector3d p2r = compensateRotation(p2, nrot, rotCenter);
-                points.push_back(p2r);
-                markers.push_back(p2r);
-                colorindex.push_back(0);
-                points.push_back(rnext);
-                markers.push_back(rnext);
-                colorindex.push_back(1);
-                double q;
-                if (cmd.has("Q")) {
-                    q = cmd.getValue("Q");
-                    if (q>0) {
-                        Base::Vector3d temp(next);
-                        for(temp.*pz=r;temp.*pz>next.*pz;temp.*pz-=q) {
-                            Base::Vector3d pr = compensateRotation(temp, nrot, rotCenter);
-                            markers.push_back(pr);
-                        }
-                    }
-                }
-                Base::Vector3d p3(next);
-                p3.*pz = last.*pz;
-                Base::Vector3d p3r = compensateRotation(p3, nrot, rotCenter);
-                points.push_back(p3r);
-                markers.push_back(p2r);
-                colorindex.push_back(0);
-                command2Edge[i] = edgeIndices.size();
-                edgeIndices.push_back(points.size());
-                edge2Command.push_back(i);
-
-                last = p3;
-                A = a;
-                B = b;
-                C = c;
-                lrot = nrot;
-
-
-            } else if ((name=="G38.2")||(name=="38.3")||(name=="G38.4")||(name=="G38.5")){
-                // Straight probe
-                Base::Vector3d p1(next.x,next.y,last.z);
-                points.push_back(p1);
-                colorindex.push_back(0);
-                points.push_back(next);
-                colorindex.push_back(2);
-                Base::Vector3d p3(next.x,next.y,last.z);
-                points.push_back(p3);
-                colorindex.push_back(0);
-                command2Edge[i] = edgeIndices.size();
-                edgeIndices.push_back(points.size());
-                edge2Command.push_back(i);
-            } else if(name=="G17") {
-                pz = &Base::Vector3d::z;
-            } else if(name=="G18") {
-                pz = &Base::Vector3d::y;
-            } else if(name=="G19") {
-                pz = &Base::Vector3d::x;
-            }
-        }
 
         if (!edgeIndices.empty()) {
             pcLineCoords->point.setNum(points.size());
